wrapper message ext encoder
